api/requirements.txt
Django==5.0.6
psycopg[binary]==3.1.19
django-environ==0.11.2
djangorestframework==3.15.2
uvicorn==0.30.3
gunicorn==22.0.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0